backend/requirements.txt
alembic==1.16.5
bcrypt==5.0.0
blinker==1.9.0
click==8.3.0
Faker==37.8.0
Flask==3.0.0
Flask-Bcrypt==1.0.1
Flask-Cors==4.0.0
Flask-JWT-Extended==4.6.0
Flask-Migrate==4.0.5
Flask-SQLAlchemy==3.1.1
gunicorn==21.2.0
itsdangerous==2.2.0
Jinja2==3.1.6
Mako==1.3.10
MarkupSafe==3.0.3
packaging==25.0
psycopg2-binary==2.9.10
PyJWT==2.10.1
python-dotenv==1.0.0
SQLAlchemy==2.0.43
sqlalchemy-serializer==1.4.22
typing_extensions==4.15.0
tzdata==2025.2
Werkzeug==3.1.3
